updated layout to follow proper structure
