+++ src/app/(layout)/blog/layout.tsx
