+++ src/app/(layout)/contact/page.tsx
@@ -1,0 +1,17 @@
+import PageBanner from '../../components/PageBanner';
+import PageBannerTitle from '../../components/PageBannerTitle';
+import ContentForm from '../../components/ContentForm';
+import Section from '../../components/Section';
+
+export default function Page() {
+  return (
+    <>
+      <PageBanner>
+        <PageBannerTitle>Contact Us</PageBannerTitle>
+      </PageBanner>
+      <Section>
+        <ContentForm />
+      </Section>
+    </>
+  );
+}
