+++ src/app/(layout)/page.tsx
@@ -1,19 +1,18 @@
-import HomepageBanner from './components/HomepageBanner';
-import Section from './components/Section';
-import Workflow from './components/Workflow';
-import Features from './components/Features';
-import Faqs from './components/Faqs';
-import Subscribe from './components/Subscribe';
-import Services from './components/Services';
-import { Testimonial } from './components/Testimonial';
-import Sponsor from './components/Sponsor';
-import BlogArticle from './components/BlogArticle';
-import About from './components/About';
-import AppLayout from './layout/layout';
+import HomepageBanner from '../components/HomepageBanner';
+import Section from '../components/Section';
+import Workflow from '../components/Workflow';
+import Features from '../components/Features';
+import Faqs from '../components/Faqs';
+import Subscribe from '../components/Subscribe';
+import Services from '../components/Services';
+import { Testimonial } from '../components/Testimonial';
+import Sponsor from '../components/Sponsor';
+import BlogArticle from '../components/BlogArticle';
+import About from '../components/About';
 
 const Home = () => {
   return (
-    <AppLayout>
+    <>
       <HomepageBanner />
       <Section>
         <About />
@@ -42,7 +41,7 @@ const Home = () => {
       <Section style='mb-0 lg:mb-0 lg:mt-40'>
         <Sponsor />
       </Section>
-    </AppLayout>
+    </>
   );
 };
