--- src/app/blog/page.tsx
@@ -1,19 +1,0 @@
-import PageBanner from "../components/PageBanner";
-import PageBannerTitle from "../components/PageBannerTitle";
-import Blogs from "../components/Blogs";
-import Section from "../components/Section";
-import AppLayout from '../layout/layout';
-
-export default function Page() {
-  return (
-    <AppLayout>
-      <PageBanner>
-        <PageBannerTitle>Latest Article</PageBannerTitle>
-      </PageBanner>
-
-      <Section>
-        <Blogs />
-      </Section>
-    </AppLayout>
-  );
-}
